zl1: port a few synaptics driver updates
 - input: synaptics_dsx: add update bounds checks.
 - input: synaptics_dsx: add bounds checks for firmware id
 - input: synaptics_dsx: Save fw size to fwu->image_size

+++ drivers/input/touchscreen/synaptics_dsx_letv/synaptics_dsx_gesture.c
@@ -878,7 +878,7 @@ static ssize_t udg_sysfs_trace_data_show(struct file *data_file,
 	if (count < trace_data_size) {
 		dev_err(rmi4_data->pdev->dev.parent,
 				"%s: Not enough space (%d bytes) in buffer\n",
-				__func__, count);
+				__func__, (unsigned int) count);
 		return -EINVAL;
 	}
 
@@ -955,7 +955,7 @@ static ssize_t udg_sysfs_template_data_show(struct file *data_file,
 	if (count < udg->template_data_size) {
 		dev_err(rmi4_data->pdev->dev.parent,
 				"%s: Not enough space (%d bytes) in buffer\n",
-				__func__, count);
+				__func__, (unsigned int) count);
 		return -EINVAL;
 	}
